+++ src/Math/Algebra/Hspray.hs
@@ -37,8 +37,6 @@ module Math.Algebra.Hspray
   , leadingTerm
   , sprayDivision
   , groebner
-  , groebner2
-  , groebner3
   ) where
 import qualified Algebra.Additive              as AlgAdd
 import qualified Algebra.Module                as AlgMod
@@ -309,7 +307,7 @@ quotient (powsQ, coeffQ) (powsP, coeffP) = (pows, coeff)
     pows = Powers expnts n
     coeff = coeffQ / coeffP
 
--- spray from monomial
+-- | spray from monomial
 fromMonomial :: Monomial a -> Spray a
 fromMonomial (pows, coeff) = HM.singleton pows coeff
 
@@ -363,8 +361,9 @@ sPolynomial p q = wp ^*^ p ^-^ wq ^*^ q
     wp = fromMonomial $ (Powers betaP n, 1 / lcoefP)
     wq = fromMonomial $ (Powers betaQ n, 1 / lcoefQ)
 
-groebner :: forall a. (Fractional a, Eq a, AlgRing.C a) => [Spray a] -> [Spray a]
-groebner sprays = go 0 j0 combins0 sprays HM.empty
+-- | Groebner basis, not minimal and not reduced
+groebner00 :: forall a. (Fractional a, Eq a, AlgRing.C a) => [Spray a] -> [Spray a]
+groebner00 sprays = go 0 j0 combins0 sprays HM.empty
   where
     j0 = length sprays
     combins0 = combn2 j0
@@ -378,28 +377,24 @@ groebner sprays = go 0 j0 combins0 sprays HM.empty
           ssnew = HM.singleton combin sfg
           spolys' = HM.union ssnew spolys
           sbarfg = sprayDivision sfg gpolys
-          -- iszero = sbarfg == AlgAdd.zero
           (i', j', gpolys', combins') = if sbarfg == AlgAdd.zero
             then
               (i+1, j, gpolys, combins)
             else
               (0, j+1, gpolys ++ [sbarfg], combn2 (j+1) \\ HM.keys spolys')
---          i' = if iszero then i+1 else 0
---          gpolys' = if iszero then gpolys else gpolys ++ [sbarfg]
---          j' = if iszero then j else j+1
---          combins' = if iszero then combins else combn2 (j+1) \\ HM.keys spolys'
 
-groebner2 :: forall a. (Fractional a, Eq a, AlgRing.C a) => [Spray a] -> [Spray a]
-groebner2 sprays = [normalize $ basis0 !! k | k <- [0 .. n-1] \\ discard]
+-- | Groebner basis, minimal but not reduced
+groebner0 :: forall a. (Fractional a, Eq a, AlgRing.C a) => [Spray a] -> [Spray a]
+groebner0 sprays = [normalize $ basis00 !! k | k <- [0 .. n-1] \\ discard]
   where
-    basis0 = groebner sprays
-    n = length basis0
+    basis00 = groebner00 sprays
+    n = length basis00
     go :: Int -> [Int] -> [Int]
     go i toRemove
       | i == n = toRemove
       | otherwise = go (i+1) toRemove'
         where
-          ltf = leadingTerm (basis0 !! i)
+          ltf = leadingTerm (basis00 !! i)
           toDrop = toRemove ++ [i]
           igo :: Int -> Bool
           igo j 
@@ -407,7 +402,7 @@ groebner2 sprays = [normalize $ basis0 !! k | k <- [0 .. n-1] \\ discard]
             | j `elem` toDrop = igo (j+1)
             | otherwise = ok || igo (j+1)
               where 
-                ok = divides (leadingTerm (basis0 !! j)) ltf
+                ok = divides (leadingTerm (basis00 !! j)) ltf
           toRemove' = if igo 0 then toDrop else toRemove
     discard = go 0 []
     normalize :: Spray a -> Spray a
@@ -415,25 +410,13 @@ groebner2 sprays = [normalize $ basis0 !! k | k <- [0 .. n-1] \\ discard]
       where
         (_, coef) = leadingTerm spray
 
-groebner3 :: forall a. (Fractional a, Eq a, AlgRing.C a) => [Spray a] -> [Spray a]
-groebner3 sprays = map reduction [0 .. n-1]
+-- | Groebner basis (minimal and reduced)
+groebner :: forall a. (Fractional a, Eq a, AlgRing.C a) => [Spray a] -> [Spray a]
+groebner sprays = map reduction [0 .. n-1]
   where
-    basis1 = groebner2 sprays
-    n = length basis1
+    basis0 = groebner0 sprays
+    n = length basis0
     reduction :: Int -> Spray a
-    reduction i = sprayDivision (basis1 !! i) rest
+    reduction i = sprayDivision (basis0 !! i) rest
       where
-        rest = [basis1 !! k | k <- [0 .. n-1] \\ [i]]
-
-
-
-      
-
-
-
-
-
-
-
-
-
+        rest = [basis0 !! k | k <- [0 .. n-1] \\ [i]]